centennial/requirements.txt
coverage==4.5.2
Django==2.1.5
django-bootstrap-modal-forms==1.3.1
Pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pytz==2018.9
